Implement a solver for the water jug puzzle
Let X and Y be the volumes of the jugs and Z the
volume we want to reach to.
The riddle can then be seen as the solution to a
Diophantine equation of the form pX + qY = Z.

`p` and `q` indicate how many times a jug needs
to be filled (if positive) or emptied (if negative).

The above equation is solvable iff gcd(X, Y) divides Z.

To find the sequence of operations, the following
algorithm is applied:
 * Repeat until the desired amount of water
   is obtained:
     – Fill one jug.
     – Transfer water into the other jug.
       Whenever the second jug becomes full,
       empty it out.

In order to know which sequence of actions is
optimal (i.e. less number of actions taken),
we will consider both possible scenarios:
 1-  Always pour from jug 1 into jug 2
 2-  Always pour from jug 2 into jug 1
and check which reaches the solution in the
minimum number of steps